+++ app/Http/Controllers/UserController.php
@@ -7,13 +7,12 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-
 namespace App\Http\Controllers;
 
 use App\Http\Requests;
 use App\User;
 use Illuminate\Http\Request;
-
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 /**
  * Class UserController
  *
@@ -27,30 +26,55 @@ class UserController extends Controller
      * Display a listing of the resource.
      *
      * @author MS
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function index()
     {
-        $user = User::latest()->get();
-        return view('user.index', compact('user'));
+        $messages = $this->getMessages();
+        $user = null;
+
+        try {
+
+            $user = User::query();
+
+            if (!empty($filter = $this->getTableFilter())) {
+                foreach ($filter as $field => $query) {
+
+                    $user->where($field, 'like', '%' . $query . '%');
+                }
+                if (!$user->count()) {
+                    $messages['info'] = 'No records were found that matched your filter';
+                }
+            }
+
+            $user = $user->paginate($this->getPageLimit());
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError('Error occurred getting locations: ' . $e->getMessage());
+            $messages['error'] = 'Error occurred getting locations';
+
+        }
+
+        return View('user.index', ['user' => $user, 'messages' => $messages]);
     }
 
     /**
      * Show the form for creating a new resource.
      *
      * @author MS
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function create()
     {
-        return view('user.create');
+        return view('user.create', ['messages' => $this->getMessages()]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @author MS
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function store(Request $request)
     {
@@ -58,13 +82,24 @@ class UserController extends Controller
             'name' => 'required',
             'email' => 'required|email',
             'password' => 'required',
-            'merchant' => 'required',
+            'merchant_id' => 'required',
         ]);
-        $array = $request->all();
 
-        $array['password'] = bcrypt($array['password']);
-        User::create($array);
-        return redirect('user');
+        $message = ['success','New User has been successfully created'];
+
+        try {
+
+            $array = $request->all();
+            $array['password'] = bcrypt($array['password']);
+            User::create($array);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError('Could not successfully create new User' . $e->getMessage());
+            $message = ['error','Could not successfully create new User'];
+        }
+
+        return redirect('user')->with($message[0], $message[1]);
     }
 
     /**
@@ -72,12 +107,26 @@ class UserController extends Controller
      *
      * @author MS
      * @param  int  $id
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function show($id)
     {
-        $user = User::findOrFail($id);
-        return view('user.show', compact('user'));
+        $user = null;
+        $messages = $this->getMessages();
+
+        try {
+
+            $user = User::findOrFail($id);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError(
+                'Could not find user with ID: [' . $id . ']; User does not exist: ' . $e->getMessage()
+            );
+            $messages['error'] = 'Could not find user with ID: [' . $id . ']; User does not exist';
+        }
+
+        return view('user.show', ['user' => $user, 'messages' => $messages]);
     }
 
     /**
@@ -85,12 +134,26 @@ class UserController extends Controller
      *
      * @author MS
      * @param  int  $id
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function edit($id)
     {
-        $user = User::findOrFail($id);
-        return view('user.edit', compact('user'));
+        $user = null;
+        $messages = $this->getMessages();
+
+        try {
+
+            $user = User::findOrFail($id);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError(
+                'Could not get user with ID [' . $id . '] for editing; User does not exist:' . $e->getMessage()
+            );
+            $messages['error'] = 'Could not get user with ID [' . $id . '] for editing; User does not exist';
+        }
+
+        return view('user.edit', ['user' => $user, 'messages' => $messages]);
     }
 
     /**
@@ -99,7 +162,7 @@ class UserController extends Controller
      * @author MS
      * @param  int  $id
      * @param Request $request
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function update($id, Request $request)
     {
@@ -108,13 +171,24 @@ class UserController extends Controller
             'password' => 'required',
         ]);
 
-        $user = User::findOrFail($id);
+        $message = ['success', 'User details were successfully updated'];
 
-        $input = $request->all();
+        try {
 
-        $input['password'] = bcrypt($input['password']);
-        $user->update($input);
-        return redirect('user');
+            $user = User::findOrFail($id);
+            $input = $request->all();
+            $input['password'] = bcrypt($input['password']);
+            $user->update($input);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError(
+                'Could not update User with ID [' . $id . ']; User does not exist' . $e->getMessage()
+            );
+            $message = ['error', 'Could not update User with ID [' . $id . ']; User does not exist'];
+        }
+
+        return redirect()->back()->with($message[0], $message[1]);
     }
 
     /**
@@ -122,12 +196,50 @@ class UserController extends Controller
      *
      * @author MS
      * @param  int  $id
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function destroy($id)
     {
-        User::destroy($id);
-        return redirect('user');
+        $message = ['success','User was successfully deleted'];
+        try {
+
+            User::destroy($id);
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError('Deletion of this record did not complete successfully' . $e->getMessage());
+            $message = ['error', 'Deletion of this record did not complete successfully'];
+        }
+
+        return redirect('user')->with($message[0], $message[1]);
     }
 
+    /**
+     * Delete
+     *
+     * @author MS
+     * @param int $id
+     * @return \Illuminate\View\View
+     */
+    public function delete($id)
+    {
+        $user = null;
+        $messages = $this->getMessages();
+
+        try {
+
+            $user = User::findOrFail($id);
+            $user->type = 'user';
+            $user->controller = 'User';
+
+        } catch (ModelNotFoundException $e) {
+
+            $this->logError(
+                'Could not get user with ID: [' . $id . ']; User does not exist: ' . $e->getMessage()
+            );
+            $messages['error'] = 'Could not get user with ID: [' . $id . ']; User does not exist';
+        }
+
+        return view('includes.page.confirm_delete', ['object' => $user, 'messages' => $messages]);
+    }
 }