+++ resources/views/installations/index.blade.php
@@ -49,7 +49,9 @@
 
                     {{-- ACTION BUTTONS --}}
                     <td class="col-xs-3 col-sm-2 col-md-2 col-lg-1 text-right">
-                        @include('includes.form.record_buttons_edit_only', ['record' => $item, 'crudName' => 'installations'])
+                        @include('includes.form.record_actions', ['id' => $item->id,
+                            'actions' => ['edit' => 'Edit']
+                        ])
                     </td>
                 </tr>
             @endforeach